requirements.txt
Flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
PyYAML==6.0.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.3
PyJWT==2.8.0
